Move imports from graphs to plots
