+++ src/c++/perf_analyzer/genai-perf/genai_perf/graphs/plot_manager.py
@@ -26,10 +26,10 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
-from genai_perf.graphs.box_plot import BoxPlot
-from genai_perf.graphs.heat_map import HeatMap
-from genai_perf.graphs.scatter_plot import ScatterPlot
 from genai_perf.llm_metrics import Statistics
+from genai_perf.plots.box_plot import BoxPlot
+from genai_perf.plots.heat_map import HeatMap
+from genai_perf.plots.scatter_plot import ScatterPlot
 from genai_perf.utils import scale
